fix return type

+++ pracice-1-sample/src/main/java/practice1/Calc.java
@@ -5,7 +5,7 @@ public class Calc {
         return a + b;
     }
 
-    public double div(int a, int b) {
+    public int div(int a, int b) {
         return a / b;
     }
 }
